Fix circular imports by do_naive_bayes_prediction method
