+++ src/skmultiflow/bayes/__init__.py
@@ -2,6 +2,7 @@
 The :mod:`skmultiflow.bayes` module includes Bayes learning methods.
 """
 
+from .utils import do_naive_bayes_prediction
 from .naive_bayes import NaiveBayes
 
-__all__ = ["NaiveBayes"]
+__all__ = ["do_naive_bayes_prediction", "NaiveBayes"]
