+++ src/skmultiflow/trees/hoeffding_anytime_tree.py
@@ -6,7 +6,7 @@ from skmultiflow.trees.hoeffding_tree import HoeffdingTree
 from skmultiflow.trees.info_gain_split_criterion import InfoGainSplitCriterion
 from skmultiflow.trees.nominal_attribute_class_observer import NominalAttributeClassObserver
 from skmultiflow.trees.numeric_attribute_class_observer_gaussian import NumericAttributeClassObserverGaussian
-from skmultiflow.trees.utils import do_naive_bayes_prediction
+from skmultiflow.bayes import do_naive_bayes_prediction
 from skmultiflow.utils.utils import get_dimensions, calculate_object_size
 
 Node = HoeffdingTree.Node
