+++ src/skmultiflow/trees/lc_hoeffding_tree.py
@@ -2,7 +2,7 @@ from skmultiflow.trees.hoeffding_tree import HoeffdingTree
 from skmultiflow.trees.numeric_attribute_class_observer_gaussian import NumericAttributeClassObserverGaussian
 from skmultiflow.trees.nominal_attribute_class_observer import NominalAttributeClassObserver
 from skmultiflow.utils.utils import *
-from skmultiflow.bayes.naive_bayes import do_naive_bayes_prediction
+from skmultiflow.bayes import do_naive_bayes_prediction
 
 GINI_SPLIT = 'gini'
 INFO_GAIN_SPLIT = 'info_gain'
